Vendor robin_map as demanded by nanobind
While we could have used robin-map-devel as it is shipped in Fedora we
decided to package the exact version requirement of nanobind in
nanobind's directory structure. This way we avoid future problems with
version conflicts. Also, by using the vendored robin_map we can ship to
f40 which wouldn't be possible unless we update robin-map-devel to at
least v1.3.0. That version is required by nanobind.